# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Course:
    description: Course defines the structure of a course
    properties:
      code:
        description: Code for the course
        type: string
        x-go-name: Code
      course_credits:
        description: Number of credits for the course
        format: int64
        minimum: 1
        type: integer
        x-go-name: Credits
      grade:
        description: Grade received by the student
        type: string
        x-go-name: Grade
      name:
        description: Name of the course
        type: string
        x-go-name: Name
    required:
    - name
    - code
    - course_credits
    - grade
    type: object
    x-go-package: github.com/killtheverse/nitd-results/app/models
  ErrorResponse:
    description: Response for returning an error message
    properties:
      content:
        description: Content to displayed along with the message
        type: object
        x-go-name: Content
      message:
        description: Error message
        type: string
        x-go-name: Message
      status:
        description: Status of the response
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: github.com/killtheverse/nitd-results/app/models
  PaginatedResponse:
    description: Response for returning a list of students
    properties:
      count:
        description: Count of students in response
        format: int64
        maximum: 100
        minimum: 0
        type: integer
        x-go-name: Count
      message:
        description: Message in the response
        type: string
        x-go-name: Message
      next:
        description: Path to fetch the students next in the list
        type: string
        x-go-name: Next
      previous:
        description: Path to fetch the students before the items currently in the
          list
        type: string
        x-go-name: Previous
      results:
        description: |-
          List of students
          collection format: Student
        items:
          $ref: '#/definitions/Student'
        type: array
        x-go-name: Students
      status:
        description: Status of the response
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: github.com/killtheverse/nitd-results/app/models
  Response:
    description: Response for returning a single student
    properties:
      message:
        description: Message in the response
        type: string
        x-go-name: Message
      status:
        description: Status of the response
        format: int64
        type: integer
        x-go-name: Status
      student:
        $ref: '#/definitions/Student'
    type: object
    x-go-package: github.com/killtheverse/nitd-results/app/models
  Semester:
    description: Semester defines the structure of a semester
    properties:
      cgpa:
        description: CGPA after this semester
        format: float
        type: number
        x-go-name: CGPA
      courses:
        description: Information about courses taken by the student in the semester
        items:
          $ref: '#/definitions/Course'
        type: array
        x-go-name: Courses
      earned_credits:
        description: Credits earned by the student in that semester
        format: int64
        minimum: 1
        type: integer
        x-go-name: Credits
      number:
        description: Semester number
        format: int64
        type: integer
        x-go-name: Number
      sgpa:
        description: SGPA earned for the semester by the student
        format: float
        type: number
        x-go-name: SGPA
    required:
    - number
    - earned_credits
    - sgpa
    - cgpa
    - courses
    type: object
    x-go-package: github.com/killtheverse/nitd-results/app/models
  Student:
    description: Student defines the structure of a student
    properties:
      branch:
        description: Branch of the student
        type: string
        x-go-name: Branch
      cgpa:
        description: Current CGPA of the student
        format: float
        type: number
        x-go-name: CGPA
      created_at:
        description: Time at which the instance was created in the database
        format: date-time
        type: string
        x-go-name: CreatedAt
      name:
        description: Name of the student
        type: string
        x-go-name: Name
      program:
        description: 'Program in which the student is enrolled in '
        type: string
        x-go-name: Program
      roll_no:
        description: Roll number of student
        type: string
        x-go-name: Roll
      semesters:
        description: Information about the previous semesters of a student
        items:
          $ref: '#/definitions/Semester'
        type: array
        x-go-name: Semesters
      updated_at:
        description: Time at which the student instance was last updated in the database
        format: date-time
        type: string
        x-go-name: UpdatedAt
    required:
    - roll_no
    - name
    - program
    - branch
    - cgpa
    - semesters
    type: object
    x-go-package: github.com/killtheverse/nitd-results/app/models
info:
  contact:
    email: r.dev2000@gmail.com
    name: Rahul Dev Kureel
  description: Documentation for NITD Results API
  title: for NITD Results API
  version: 1.0.0
paths:
  /students:
    get:
      consumes:
      - application/json
      description: Returns a list of students filtered by parameters
      operationId: listStudents
      parameters:
      - description: The branch of the students
        in: query
        name: branch
        type: string
        x-go-name: Branch
      - description: The program of the students
        in: query
        name: program
        type: string
        x-go-name: Program
      - description: Limit on the number of results
        format: int64
        in: query
        name: limit
        type: integer
        x-go-name: Limit
      - description: Offset on the results
        format: int64
        in: query
        name: offset
        type: integer
        x-go-name: Offset
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/PaginatedResponse'
        default:
          $ref: '#/responses/ErrorResponse'
      schemes:
      - http
      tags:
      - students
  /students/{roll_number}:
    get:
      consumes:
      - application/json
      description: Returns information about a particular student
      operationId: studentDetail
      parameters:
      - description: The roll number of the student
        in: path
        name: roll_number
        required: true
        type: string
        x-go-name: Roll
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/Response'
        default:
          $ref: '#/responses/ErrorResponse'
      schemes:
      - http
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Updates a student if it exists in the database, otherwise creates
        a new entry
      operationId: updateStudent
      parameters:
      - description: The roll number of the student
        in: path
        name: roll_number
        required: true
        type: string
        x-go-name: Roll
      produces:
      - application/json
      responses:
        "201":
          $ref: '#/responses/Response'
        default:
          $ref: '#/responses/ErrorResponse'
      schemes:
      - http
      security:
      - bearerAuth:
        - '[]'
      tags:
      - students
produces:
- application/json
responses:
  ErrorResponse:
    description: ""
    schema:
      $ref: '#/definitions/ErrorResponse'
  PaginatedResponse:
    description: ""
    schema:
      $ref: '#/definitions/PaginatedResponse'
  Response:
    description: ""
    schema:
      $ref: '#/definitions/Response'
schemes:
- http
swagger: "2.0"